+++ python/lsst/ts/wep/task/calcZernikesTask.py
@@ -27,7 +27,6 @@ __all__ = [
 
 import abc
 from itertools import zip_longest
-
 import astropy.units as u
 import lsst.pex.config as pexConfig
 import lsst.pipe.base as pipeBase
@@ -39,14 +38,11 @@ from lsst.pipe.base import (
     QuantumContext,
     connectionTypes,
 )
-from lsst.ts.wep.utils.enumUtils import DefocalType
 from lsst.ts.wep.task.combineZernikesSigmaClipTask import CombineZernikesSigmaClipTask
 from lsst.ts.wep.task.donutStamps import DonutStamps
 from lsst.ts.wep.task.donutStampSelectorTask import DonutStampSelectorTask
 from lsst.ts.wep.task.estimateZernikesTieTask import EstimateZernikesTieTask
-from lsst.ts.wep.task.fitDonutRadiusTask import FitDonutRadiusTask
 from lsst.utils.timer import timeMethod
-from lsst.ts.wep.utils import getTaskInstrument
 
 pos2f_dtype = np.dtype([("x", "<f4"), ("y", "<f4")])
 
@@ -116,12 +112,14 @@ class CalcZernikesTaskConfig(
         target=DonutStampSelectorTask,
         doc="How to select donut stamps.",
     )
-    fitDonutRadius = pexConfig.ConfigurableField(
-        target=FitDonutRadiusTask,
-        doc="How to estimate donut radius.",
-    )
     doDonutStampSelector = pexConfig.Field(
-        doc="Whether or not to run donut stamp selector.",
+        doc="Whether or not to run donut stamp selector."
+        + "If this is False, then we do not get donutQualityTable."
+        + "(The default is True). It is also possible to run"
+        + "donut stamp selector (with this config set to True), but"
+        + "turn off doSelection config inside the donut stamp selector,"
+        + "which would return all donuts as selected, as well as"
+        + "returning a quality table.",
         dtype=bool,
         default=True,
     )
@@ -151,10 +149,11 @@ class CalcZernikesTask(pipeBase.PipelineTask, metaclass=abc.ABCMeta):
         self.donutStampSelector = self.config.donutStampSelector
         self.makeSubtask("donutStampSelector")
 
-        self.fitDonutRadius = self.config.fitDonutRadius
-        self.makeSubtask("fitDonutRadius")
-
         self.doDonutStampSelector = self.config.doDonutStampSelector
+
+        # Initialize the donut stamps to None
+        self.stampsExtra = None
+        self.stampsIntra = None
 
     def initZkTable(self) -> QTable:
         """Initialize the table to store the Zernike coefficients
@@ -315,7 +314,13 @@ class CalcZernikesTask(pipeBase.PipelineTask, metaclass=abc.ABCMeta):
                     * u.pixel
                 )
             )
-            for key in ["MAG", "SN", "ENTROPY", "FRAC_BAD_PIX", "MAX_POWER_GRAD"]:
+            for key in [
+                "MAG",
+                "SN",
+                "ENTROPY",
+                "FRAC_BAD_PIX",
+                "MAX_POWER_GRAD"
+            ]:
                 for stamps, foc in [
                     (intraStamps, "intra"),
                     (extraStamps, "extra"),
@@ -327,45 +332,66 @@ class CalcZernikesTask(pipeBase.PipelineTask, metaclass=abc.ABCMeta):
                     row[f"{foc}_{key.lower()}"] = val
             zkTable.add_row(row)
 
-        zkTable.meta["intra"] = {}
-        zkTable.meta["extra"] = {}
-
-        for dict_, stamps in [
-            (zkTable.meta["intra"], intraStamps),
-            (zkTable.meta["extra"], extraStamps),
-        ]:
-            dict_["det_name"] = "" if len(stamps) == 0 else stamps.metadata["DET_NAME"]
-            dict_["visit"] = "" if len(stamps) == 0 else stamps.metadata["VISIT"]
-            dict_["dfc_dist"] = "" if len(stamps) == 0 else stamps.metadata["DFC_DIST"]
-            dict_["band"] = "" if len(stamps) == 0 else stamps.metadata["BANDPASS"]
-            dict_["boresight_rot_angle_rad"] = (
-                "" if len(stamps) == 0 else stamps.metadata["BORESIGHT_ROT_ANGLE_RAD"]
-            )
-            dict_["boresight_par_angle_rad"] = (
-                "" if len(stamps) == 0 else stamps.metadata["BORESIGHT_PAR_ANGLE_RAD"]
-            )
-            dict_["boresight_alt_rad"] = (
-                "" if len(stamps) == 0 else stamps.metadata["BORESIGHT_ALT_RAD"]
-            )
-            dict_["boresight_az_rad"] = (
-                "" if len(stamps) == 0 else stamps.metadata["BORESIGHT_AZ_RAD"]
-            )
-            dict_["boresight_ra_rad"] = (
-                "" if len(stamps) == 0 else stamps.metadata["BORESIGHT_RA_RAD"]
-            )
-            dict_["boresight_dec_rad"] = (
-                "" if len(stamps) == 0 else stamps.metadata["BORESIGHT_DEC_RAD"]
-            )
-            dict_["mjd"] = "" if len(stamps) == 0 else stamps.metadata["MJD"]
-
-        if len(intraStamps) > 0:
-            zkTable.meta["cam_name"] = intraStamps.metadata["CAM_NAME"]
-        else:
-            zkTable.meta["cam_name"] = extraStamps.metadata["CAM_NAME"]
-        if len(intraStamps) > 0 and len(extraStamps) > 0:
-            assert intraStamps.metadata["CAM_NAME"] == extraStamps.metadata["CAM_NAME"]
+        zkTable.meta = self.createZkTableMetadata()
 
         return zkTable
+
+    def createZkTableMetadata(self):
+        """Create the metadata for the Zernike table.
+
+        Returns
+        -------
+        metadata : dict
+            Metadata for the Zernike table
+        """
+        meta = {}
+        meta["intra"] = {}
+        meta["extra"] = {}
+        cam_name = None
+
+        if self.stampsIntra is None and self.stampsExtra is None:
+            raise ValueError(
+                "No stamps available. Cannot create metadata."
+            )
+
+        for dict_, stamps in [
+            (meta["intra"], self.stampsIntra),
+            (meta["extra"], self.stampsExtra),
+        ]:
+            if stamps is None:
+                continue
+            dict_["det_name"] = stamps.metadata["DET_NAME"]
+            dict_["visit"] = stamps.metadata["VISIT"]
+            dict_["dfc_dist"] =stamps.metadata["DFC_DIST"]
+            dict_["band"] = stamps.metadata["BANDPASS"]
+            dict_["boresight_rot_angle_rad"] = (
+                stamps.metadata["BORESIGHT_ROT_ANGLE_RAD"]
+            )
+            dict_["boresight_par_angle_rad"] = (
+                stamps.metadata["BORESIGHT_PAR_ANGLE_RAD"]
+            )
+            dict_["boresight_alt_rad"] = (
+               stamps.metadata["BORESIGHT_ALT_RAD"]
+            )
+            dict_["boresight_az_rad"] = (
+                 stamps.metadata["BORESIGHT_AZ_RAD"]
+            )
+            dict_["boresight_ra_rad"] = (
+                stamps.metadata["BORESIGHT_RA_RAD"]
+            )
+            dict_["boresight_dec_rad"] = (
+                stamps.metadata["BORESIGHT_DEC_RAD"]
+            )
+            dict_["mjd"] = stamps.metadata["MJD"]
+            if cam_name is None:
+                cam_name = stamps.metadata["CAM_NAME"]
+
+        meta["cam_name"] = cam_name
+
+        if self.stampsIntra is not None and self.stampsExtra is not None:
+            assert self.stampsIntra.metadata["CAM_NAME"] == self.stampsExtra.metadata["CAM_NAME"]
+
+        return meta
 
     def empty(self, qualityTable=None, zernikeTable=None) -> pipeBase.Struct:
         """Return empty results if no donuts are available. If
@@ -392,6 +418,9 @@ class CalcZernikesTask(pipeBase.PipelineTask, metaclass=abc.ABCMeta):
             "ENTROPY_SELECT",
             "SN_SELECT",
             "FINAL_SELECT",
+            "RADIUS",
+            "X_PIX_LEFT_EDGE",
+            "X_PIX_RIGHT_EDGE",
             "DEFOCAL_TYPE",
         ]
         if qualityTable is None:
@@ -401,6 +430,7 @@ class CalcZernikesTask(pipeBase.PipelineTask, metaclass=abc.ABCMeta):
 
         if zernikeTable is None:
             zkTable = self.initZkTable()
+            zkTable.meta = self.createZkTableMetadata()
         else:
             zkTable = zernikeTable
 
@@ -421,121 +451,6 @@ class CalcZernikesTask(pipeBase.PipelineTask, metaclass=abc.ABCMeta):
         outputs = self.run(**inputs, numCores=butlerQC.resources.num_cores)
         butlerQC.put(outputs, outputRefs)
 
-    def getZ4FromDonutRadius(self, donutStampsExtra, donutStampsIntra, qualityTable=None):
-        """Get the Z4 from the donut radius and return a zkTable with the avg
-        Zernike coefficients with the computed Z4 and all other Zernikes 0.
-        This is used when no donuts are available or when the donut stamp
-        selector fails to select any donuts.
-        The Z4 is computed from the median radius of the donuts that were
-        successfully fit with fitDonutRadiusTask.
-
-        Parameters
-        ----------
-        donutStampsExtra : DonutStamps
-            The extrafocal stamps
-        donutStampsIntra : DonutStamps
-            The intrafocal stamps
-        qualityTable : astropy.table.QTable
-            Quality table created with donut stamp input.
-
-        Returns
-        -------
-        struct : lsst.pipe.base.Struct
-            Struct with the Zernike coefficients from fitting the donut radius
-            and the quality table. See self.empty above for full contents.
-        """
-        if len(donutStampsExtra) == 0 and len(donutStampsIntra) == 0:
-            return self.empty(qualityTable=qualityTable)
-
-        if len(donutStampsExtra) > 0:
-            refStamp = donutStampsExtra[0]
-        else:
-            refStamp = donutStampsIntra[0]
-        camName = refStamp.cam_name
-        detectorName = refStamp.detector_name
-        instrument = getTaskInstrument(
-            camName,
-            detectorName,
-            self.estimateZernikes.config.instConfigFile,
-        )
-
-        # Fit donut radius for both intra and extra stamps
-        radii_struct = self.fitDonutRadius.run(donutStampsExtra, donutStampsIntra)
-        # Remove the fits that failed
-        valid_radii = radii_struct.donutRadiiTable[radii_struct.donutRadiiTable["FAIL_FLAG"] == 0]
-
-        if len(valid_radii) == 0:
-            self.log.info("No donuts were successfully fit. Returning empty struct.")
-            return self.empty(qualityTable=qualityTable)
-
-        n_intra = len(valid_radii[valid_radii["DFC_TYPE"] == DefocalType.Intra.value])
-        n_extra = len(valid_radii[valid_radii["DFC_TYPE"] == DefocalType.Extra.value])
-        self.log.info(f"Number of intra donuts successfully fit: {n_intra}")
-        self.log.info(f"Number of extra donuts successfully fit: {n_extra}")
-
-        # We will only use the fits of the side of focus that has more donuts.
-        dominant_type = DefocalType.Intra.value if n_intra >= n_extra else DefocalType.Extra.value
-        dominant_radii = valid_radii[valid_radii["DFC_TYPE"] == dominant_type]["RADIUS"]
-        median_radius = np.median(dominant_radii.value)
-        self.log.info(f"Median donut radius (in pixels): {median_radius}.")
-
-        # Compute the defocus offset from the median radius
-        # We multiply by sqrt(4 * f^2 - 1) * pixel_scale to get the defocus
-        # offset in meters, and subtract the nominal defocus of the wavefront
-        # sensors.
-        median_offset = (
-            median_radius * np.sqrt(4 * instrument.focalRatio**2 - 1)
-            * instrument.pixelSize
-            - instrument.defocalOffset
-        )
-        self.log.info(f"Defocus offset computed from donut radii: {median_offset*1e6} um.")
-
-        # If the dominant type is intra, the donuts are biggest in the
-        # extra-focal sensor, so we need to move in minus direction.
-        # If the dominant type is extra, the donuts are biggest in the
-        # intra-focal sensor, so we need to move in plus direction.
-        # We are assuming here that we are not so defocused that both
-        # donuts are larger than the default size. If that were the
-        # case, we would expect fitDonutRadius to fail and return an
-        # empty struct earlier.
-        if dominant_type == DefocalType.Intra.value:
-            Z4 = instrument.offsetToZ4Defocus(median_offset)
-        else:
-            Z4 = instrument.offsetToZ4Defocus(-median_offset)
-        self.log.info(f"Z4 defocus computed from fitting zernikes: {Z4}.")
-
-        # Generate zernike vector with only Z4
-        # and set all other Zernikes to 0
-        zernikes = np.zeros(len(self.nollIndices))
-        zernikes[self.nollIndices.index(4)] = Z4
-
-        zkTable = self.initZkTable()
-        zkTable.add_row(
-            {
-                "label": "average",
-                "used": True,
-                **{
-                    f"Z{j}": zernikes[i] * u.micron
-                    for i, j in enumerate(self.nollIndices)
-                },
-                "intra_field": np.nan,
-                "extra_field": np.nan,
-                "intra_centroid": np.nan,
-                "extra_centroid": np.nan,
-                "intra_mag": np.nan,
-                "extra_mag": np.nan,
-                "intra_sn": np.nan,
-                "extra_sn": np.nan,
-                "intra_entropy": np.nan,
-                "extra_entropy": np.nan,
-                "intra_frac_bad_pix": np.nan,
-                "extra_frac_bad_pix": np.nan,
-                "intra_max_power_grad": np.nan,
-                "extra_max_power_grad": np.nan,
-            }
-        )
-        return self.empty(zernikeTable=zkTable, qualityTable=qualityTable)
-
     @timeMethod
     def run(
         self,
@@ -546,23 +461,22 @@ class CalcZernikesTask(pipeBase.PipelineTask, metaclass=abc.ABCMeta):
         # If no donuts are in the donutCatalog for a set of exposures
         # in one of the sides of focus, then attempt to get the zernikes
         # by fitting the donut radius. If that fails, return empty struct.
-        if len(donutStampsExtra) == 0 or len(donutStampsIntra) == 0:
-            return self.getZ4FromDonutRadius(
-                donutStampsExtra,
-                donutStampsIntra
-            )
+        self.stampsExtra = donutStampsExtra
+        self.stampsIntra = donutStampsIntra
+        if len(self.stampsExtra) == 0 or len(self.stampsIntra) == 0:
+            return self.empty()
 
-        # Run donut stamp selection. By default all donut stamps are selected
-        # and we are provided with donut quality table.
+        # Run donut stamp selection. By default, doSelection is turned on,
+        # and we select only donuts that pass the criteria.
+        # We are always provided with donut quality table.
         if self.doDonutStampSelector:
             self.log.info("Running Donut Stamp Selector")
-            selectionExtra = self.donutStampSelector.run(donutStampsExtra)
-            selectionIntra = self.donutStampSelector.run(donutStampsIntra)
+            selectionExtra = self.donutStampSelector.run(self.stampsExtra)
+            selectionIntra = self.donutStampSelector.run(self.stampsIntra)
             donutExtraQuality = selectionExtra.donutsQuality
             donutIntraQuality = selectionIntra.donutsQuality
             selectedExtraStamps = selectionExtra.donutStampsSelect
             selectedIntraStamps = selectionIntra.donutStampsSelect
-
             donutExtraQuality["DEFOCAL_TYPE"] = "extra"
             donutIntraQuality["DEFOCAL_TYPE"] = "intra"
             donutQualityTable = vstack([donutExtraQuality, donutIntraQuality])
@@ -574,19 +488,21 @@ class CalcZernikesTask(pipeBase.PipelineTask, metaclass=abc.ABCMeta):
                 len(selectedExtraStamps) == 0
                 or len(selectedIntraStamps) == 0
             ):
-                self.log.info("No donut stamps were selected. Fitting donut radius.")
-                empty_struct = self.getZ4FromDonutRadius(
-                    donutStampsExtra,
-                    donutStampsIntra,
-                    qualityTable=donutQualityTable
-                )
-                return empty_struct
+                self.log.info("No donut stamps were selected.")
+                return self.empty(qualityTable=donutQualityTable)
         else:
+            self.log.info("Not running Donut Stamp Selector")
             donutQualityTable = QTable([])
+            selectedExtraStamps = self.stampsExtra
+            selectedIntraStamps = self.stampsIntra
+
+        # Update stampsExtra and stampsIntra with the selected donuts
+        self.stampsExtra = selectedExtraStamps
+        self.stampsIntra = selectedIntraStamps
 
         # Estimate Zernikes from the collection of selected stamps
         zkCoeffRaw = self.estimateZernikes.run(
-            selectedExtraStamps, selectedIntraStamps, numCores=numCores
+            self.stampsExtra, self.stampsIntra, numCores=numCores
         )
         zkCoeffCombined = self.combineZernikes.run(zkCoeffRaw.zernikes)
 
@@ -596,6 +512,7 @@ class CalcZernikesTask(pipeBase.PipelineTask, metaclass=abc.ABCMeta):
             zkCoeffRaw,
             zkCoeffCombined,
         )
+        zkTable.meta["estimatorInfo"] = zkCoeffRaw.wfEstInfo
 
         return pipeBase.Struct(
             outputZernikesAvg=np.atleast_2d(np.array(zkCoeffCombined.combinedZernikes)),