+++ python/lsst/ts/wep/task/donutStampSelectorTask.py
@@ -99,6 +99,12 @@ class DonutStampSelectorTaskConfig(pexConfig.Config):
             + "to be selected."
         ),
     )
+    doSelection =  pexConfig.Field(
+        dtype=bool,
+        default=True,
+        doc="Do any donut stamp selection? If this is False then we return the donut quality table,"
+        + " but allow all donut stamps to be selected.",
+    )
 
 
 class DonutStampSelectorTask(pipeBase.Task):
@@ -145,7 +151,15 @@ class DonutStampSelectorTask(pipeBase.Task):
         )
         selectedStamps._refresh_metadata()
         # Need to copy a few other fields by hand
-        for k in ["SN", "ENTROPY", "FRAC_BAD_PIX", "MAX_POWER_GRAD", "VISIT"]:
+        for k in [
+            "SN",
+            "ENTROPY",
+            "FRAC_BAD_PIX",
+            "MAX_POWER_GRAD",
+            "RADIUS",
+            "X_PIX_LEFT_EDGE",
+            "X_PIX_RIGHT_EDGE"
+        ]:
             if k in donutStamps.metadata:
                 selectedStamps.metadata[k] = np.array(
                     [
@@ -155,7 +169,7 @@ class DonutStampSelectorTask(pipeBase.Task):
                     ]
                 )
         for key, val in donutStamps.metadata.items():
-            if key.startswith("BORESIGHT") or key == "MJD":
+            if key.startswith("BORESIGHT") or key in ["MJD", "VISIT", "DFC_DIST", "DET_NAME", "BANDPASS"]:
                 selectedStamps.metadata[key] = val
             else:
                 continue
@@ -212,6 +226,20 @@ class DonutStampSelectorTask(pipeBase.Task):
             self.log.warning(
                 "selectWithEntropy==True but ENTROPY not in stamp metadata."
             )
+
+        # Collect the donut radius metric information if available
+        donutRadii = np.full(len(donutStamps), np.nan)
+        stampsLeftEdges = np.full(len(donutStamps), np.nan)
+        stampsRightEdges = np.full(len(donutStamps), np.nan)
+        if "RADIUS" in list(donutStamps.metadata):
+            fillVals = np.asarray(donutStamps.metadata.getArray("RADIUS"))
+            donutRadii[: len(fillVals)] = fillVals
+
+            fillVals = np.asarray(donutStamps.metadata.getArray("X_PIX_LEFT_EDGE"))
+            stampsLeftEdges[: len(fillVals)] = fillVals
+
+            fillVals = np.asarray(donutStamps.metadata.getArray("X_PIX_RIGHT_EDGE"))
+            stampsRightEdges[: len(fillVals)] = fillVals
 
         # By default select all donuts,  only overwritten
         # if selectWithSignalToNoise is True
@@ -288,11 +316,21 @@ class DonutStampSelectorTask(pipeBase.Task):
             )
 
         # choose only donuts that satisfy all selected conditions
-        selected = entropySelect * snSelect * fracBadPixSelect * maxPowerGradSelect
+        if self.config.doSelection:
+            selected = entropySelect * snSelect * fracBadPixSelect * maxPowerGradSelect
+            self.log.info(
+                    f"{sum(selected)} of {len(selected)} donuts "
+                    "passed combined selection criteria."
+                )
+              # make sure we don't select more than maxSelect
+            if self.config.maxSelect != -1:
+                selected[np.cumsum(selected) > self.config.maxSelect] = False
 
-        # make sure we don't select more than maxSelect
-        if self.config.maxSelect != -1:
-            selected[np.cumsum(selected) > self.config.maxSelect] = False
+        else:
+            selected = np.ones(len(donutStamps), dtype="bool")
+            self.log.info(
+                    "Donut stamp selector is off."
+                )
 
         # store information about which donuts were selected
         # use QTable even though no units at the moment in
@@ -308,6 +346,9 @@ class DonutStampSelectorTask(pipeBase.Task):
                 entropySelect,
                 fracBadPixSelect,
                 maxPowerGradSelect,
+                donutRadii,
+                stampsLeftEdges,
+                stampsRightEdges,
                 selected,
             ],
             names=[
@@ -319,10 +360,15 @@ class DonutStampSelectorTask(pipeBase.Task):
                 "ENTROPY_SELECT",
                 "FRAC_BAD_PIX_SELECT",
                 "MAX_POWER_GRAD_SELECT",
+                "RADIUS",
+                "X_PIX_LEFT_EDGE",
+                "X_PIX_RIGHT_EDGE",
                 "FINAL_SELECT",
             ],
         )
-
+        # Add all configuration used for selection criteria as metadata;
+        # that way it doesn't have to be pulled from calcZernikes config
+        donutsQuality.meta = {"DonutStampSelectorTaskConfig": self.config.toDict()}
         self.log.info("Selected %d/%d donut stamps", selected.sum(), len(donutStamps))
 
         return pipeBase.Struct(selected=selected, donutsQuality=donutsQuality)